requirements.txt
PyYAML>=6.0
python-dotenv>=1.0.0      # load .env

# VLM Clients
anthropic>=0.40.0          # Claude Opus / Sonnet
openai>=1.50.0             # GPT-4o / GPT-4o-mini

# Image Processing
Pillow>=10.0                # resize, encode ảnh

# HTTP
requests>=2.31.0            # gọi Gemma 4 qua Google AI Studio
